+++ app/resetpw/page.tsx
@@ -3,6 +3,7 @@ import AuthButtons from '@/components/Auth/AuthButtons'
 import AuthInput from '@/components/Auth/AuthInput'
 import AuthTitle from '@/components/Auth/AuthTitle'
 import Modal from '@/components/common/Modal'
+import useAuthForm from '@/hooks/useAuthForm'
 import { reset } from '@/services/auth'
 import { Card } from '@/styles/customStyle'
 import { SignupForm } from '@/types/Auth'
@@ -14,13 +15,15 @@ import { useState } from 'react'
 
 function ResetPwPage() {
   const router = useRouter()
-  const [form, setForm] = useState<
+
+  const { form, handleInput, isFormValid } = useAuthForm<
     Pick<SignupForm, 'email' | 'phoneNumber' | 'password'>
   >({
     email: '',
     phoneNumber: '',
     password: '',
   })
+
   const [modalOpen, setModalOpen] = useState<boolean>(false)
   const [modalContent, setModalContent] = useState<ModalType>({
     isError: false,
@@ -28,15 +31,6 @@ function ResetPwPage() {
     description: '',
     onBtnClick: () => {},
   })
-
-  const handleInput = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target
-    setForm((prev) => ({ ...prev, [name]: value }))
-  }
-
-  const isFormValid = Object.values(form).every(
-    (v) => v.trim() !== '',
-  )
 
   const handleSubmit = async () => {
     try {
